+++ CTextGame/Renderer/Renderer.h
@@ -1,12 +1,12 @@
 #pragma once
 
-#include "DataType.h"
+#include <Data/DataType.h>
 
 bool RendererInit();
 bool ClearBuffer();
 
 bool NullTermimation();
-void SetConsolePointer(vector2D<unsigned int> position);
+void SetConsolePointer(vector2D<unsigned short> position);
 
 
 bool CopyDataToRendBuffer(const vector2D<unsigned int> position, const char render);
